# public/swaggerApi/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OTC Forentend API",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "{schema}://192.168.10.195:8080",
            "description": "OpenApi parameters",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        }
    ],
    "paths": {
...
        "/otc/country": {
            "get": {
                "tags": [
                    "otc-common"
                ],
                "summary": "国家列表",
                "operationId": "Country",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "code": 200,
                                    "msg": "成功",
                                    "data": {
                                        "data": [
                                            {
                                                "code": "AO",
                                                "areaCode": 244,
                                                "nameEn": "Angola",
                                                "nameZh-CN": "安哥拉",
                                                "nameZh-HK": "安哥拉",
                                                "language": ""
                                            }
                                        ],
                                        "total": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/otc/wallet": {
            "get": {
                "tags": [
                    "otc-wallet"
                ],
                "summary": "钱包余额",
                "operationId": "Wallet",
                "parameters": [
                    {
                        "name": "coin",
                        "in": "query",
                        "description": "币种名称如:BTC",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "BTC",
                                "LTC"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/wallet/address": {
            "get": {
                "tags": [
                    "otc-wallet"
                ],
                "summary": "钱包地址",
                "operationId": "addressWallet",
                "parameters": [
                    {
                        "name": "coin",
                        "in": "query",
                        "description": "币种名称如:BTC",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "BTC",
                                "LTC"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/wallet/transactions": {
            "get": {
                "tags": [
                    "otc-wallet"
                ],
                "summary": "交易记录",
                "operationId": "transactionsWallet",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "当前页，默认1",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "当前页大小，默认10",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "排序字段，默认id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "排列顺序，默认desc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "类型:NetworkIn=网络转入,NetworkOut=网络转出,TradeIn=交易买入,TradeOut=交易卖出,PlatformIn=平台转入,PlatformOut=平台转出",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "NetworkIn",
                                "NetworkOut",
                                "TradeIn",
                                "TradeOut",
                                "PlatformIn",
                                "PlatformOut"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/wallet/deposits": {
            "get": {
                "tags": [
                    "otc-wallet"
                ],
                "summary": "取得所有未到账的存款列表",
                "operationId": "depositsWallet",
                "parameters": [
                    {
                        "name": "coin",
                        "in": "query",
                        "description": "币种名称如:BTC",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "BTC",
                                "LTC"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/wallet/qrcode/{address}/{amount}": {
            "get": {
                "tags": [
                    "otc-wallet"
                ],
                "summary": "钱包收款二维码图片",
                "operationId": "qrcodeWallet",
                "parameters": [
                    {
                        "name": "address",
                        "in": "query",
                        "description": "钱包地址",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "数量",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/wallet/send": {
            "post": {
                "tags": [
                    "otc-wallet"
                ],
                "summary": "发送数字货币",
                "operationId": "sendWallet",
                "parameters": [
                    {
                        "name": "address",
                        "in": "query",
                        "description": "钱包地址",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "数量",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "secondaryPassword",
                        "in": "query",
                        "description": "资金密码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "remark",
                        "in": "query",
                        "description": "备注",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        }
    }
}